Replace index.html with this index.html
I changed the background color to my liking and I tried to change the text color. Please also change the .academy to a light green (21d675). Also please change all text that is hard to see to a green or white color that is visible. The preferred greens are 1ec96d for the letscode and 21d675 for the .academy. Thanks.